+++ app/src/main/java/nextstep/shoppingcart/ui/screen/component/CartItemContainer.kt
@@ -14,18 +14,15 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import nextstep.shoppingcart.R
 import nextstep.shoppingcart.data.CartItem
 import nextstep.shoppingcart.data.Product
 import nextstep.shoppingcart.ui.theme.Gray10
@@ -137,59 +134,11 @@ private fun CartItemInfo(
             fontSize = 16.sp,
             fontWeight = FontWeight.W400,
         )
-        CartQuantitySector(
+        QuantityAdjusContainer(
             count = count,
             modifier = Modifier.fillMaxWidth(),
-            onMinusCartItem = onMinusCartItem,
-            onPlusCartItem = onPlusCartItem
-        )
-    }
-}
-
-@Composable
-private fun CartQuantitySector(
-    count: Int,
-    modifier: Modifier = Modifier,
-    onMinusCartItem: () -> Unit,
-    onPlusCartItem: () -> Unit,
-) {
-    Row(
-        modifier = modifier,
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        QuantityAdjusButton(
-            buttonTitle = stringResource(id = R.string.cart_minus_item_button),
-            onClick = onMinusCartItem,
-        )
-        Text(
-            text = count.toString(),
-            fontSize = 22.sp,
-            fontWeight = FontWeight.W400,
-        )
-        QuantityAdjusButton(
-            buttonTitle = stringResource(id = R.string.cart_plus_item_button),
-            onClick = onPlusCartItem,
-        )
-    }
-}
-
-@Composable
-private fun QuantityAdjusButton(
-    buttonTitle: String,
-    modifier: Modifier = Modifier,
-    onClick: () -> Unit,
-) {
-    IconButton(
-        modifier = modifier
-            .clip(RoundedCornerShape(4.dp))
-            .padding(4.dp),
-        onClick = onClick
-    ) {
-        Text(
-            text = buttonTitle,
-            fontSize = 16.sp,
-            fontWeight = FontWeight.W700,
+            onMinusCartItemClick = onMinusCartItem,
+            onPlusCartItemClick = onPlusCartItem
         )
     }
 }
@@ -204,31 +153,6 @@ private fun CartItemInfoPreview() {
 
         },
         onPlusCartItem = {
-
-        },
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun CartQuantitySectorPreview() {
-    CartQuantitySector(
-        count = 2,
-        onMinusCartItem = {
-
-        },
-        onPlusCartItem = {
-
-        },
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun QuantityAdjusButtonPreview() {
-    QuantityAdjusButton(
-        buttonTitle = "+",
-        onClick = {
 
         },
     )
